+++ src/main/java/com/ixayda/iam/credential/internal/ldap/LdapSubjectFormat.java
@@ -1,0 +1,9 @@
+package com.ixayda.iam.credential.internal.ldap;
+
+enum LdapSubjectFormat {
+
+	TEXT,
+
+	BINARY_BASE64URL
+
+}