+++ src/main/java/com/ixayda/iam/credential/internal/ldap/LdapTransportSecurity.java
@@ -1,0 +1,9 @@
+package com.ixayda.iam.credential.internal.ldap;
+
+enum LdapTransportSecurity {
+
+	LDAPS,
+
+	START_TLS
+
+}